Move advanced options
